エンドコンテンツの実装

+++ Base.h
@@ -13,7 +13,7 @@ public:
 
     VECTOR BasePosition;
     VECTOR targetAngle;
-    int Health = 100;
+    float Health = 100;
     VECTOR offset = VGet(0, -5, 0);
     void Rotate(VECTOR Vector);
     void Move(VECTOR Vector);
