+++ Text.h
@@ -17,8 +17,8 @@ class UIText
 {
 private:
 public:
-    void DrawTextWithSort(int xleft, int xright, std::string text, int FontHandle, int sort, int y, bool shadow, unsigned int FontColor, unsigned int shadowcolor = GetColor(50, 50, 50), int value = 0);
-    bool fadeInText(int xleft, int xright, std::string text, int FontHandle, int sort, int y, bool shadow, unsigned int FontColor, unsigned int shadowcolor = GetColor(50, 50, 50), int value = 0, float time = 1.5f);
+    void DrawTextWithSort(int xleft, int xright, std::string text, int FontHandle, int sort, int y, bool shadow, unsigned int FontColor, unsigned int shadowcolor = GetColor(50, 50, 50), float value = 0);
+    bool fadeInText(int xleft, int xright, std::string text, int FontHandle, int sort, int y, bool shadow, unsigned int FontColor, unsigned int shadowcolor = GetColor(50, 50, 50), float value = 0, float time = 1.5f);
     void resetAlpha();
 };
 void fontLoad();
